// src/lib/api/generated/model/dayAvailabilityStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Turnerito API
 * API del turnero SaaS multi-tenant
 * OpenAPI spec version: 1.0
 */

export type DayAvailabilityStatus = typeof DayAvailabilityStatus[keyof typeof DayAvailabilityStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DayAvailabilityStatus = {
  available: 'available',
  closed: 'closed',
  time_off: 'time_off',
  full: 'full',
} as const;